S2 P1 time_control fix6
Sprint 2 Part 1 time_control fix6

+++ sprint2/problems/time_control/solution/src/model.h
@@ -16,7 +16,7 @@ namespace model {
 
 using Dimension = int64_t;
 using Coord = Dimension;
-using RealDimension = float;
+using RealDimension = double;
 
 enum Direction {
     UNKNOWN = -1,
